Fetch last_modified_gmt and last_touch
as requested by @martensson in #43.

These fields are set by the LastPass servers.
They are set on the Account struct when reading via Accounts().
They are ignored for all other methods including Add() and Update()
(although Add() and Update() will update these fields on the server
side).

The LastPass CLI parses these timestamps as "unsigned long int" values
in
https://github.com/lastpass/lastpass-cli/blob/8767b5e53192ad4e72d1352db4aa9218e928cbe1/format.c#L67

unsigned long int has a minimum size of 32 bits (but could be 64 bits).

In this library, we just return the timestamp as-is as string.
The user can convert them to their needs (e.g. integer or time.Time).

lastModifiedGMT is the Unix timestamp, i.e. seconds since January 1, 1970 UTC

lastTouch means when the item was "last used".
This timestamp is not necessarily in UTC time zone.