Add(), Update(), Delete() item in shared folder
So far, reading from shared folders has been supported, but not adding,
updating, deleting items to / from shared folders.

This commit adds support to Add(), Update(), Delete() an item to / from a shared folder.

When writing an item to a shared folder, fields will be encrypted with
the shared folder's sharing key.

A descriptive error message is returned if the shared folder is read-only.

Moving an item from / to a shared folder is not supported.
Instead, the caller of this library should first Delete() and then Add()
the item. That yields the same behaviour as the LastPass CLI "lpass mv" command
since a new account ID will be generated when items are moved from / to
a shared folder:
https://github.com/lastpass/lastpass-cli/blob/8767b5e53192ad4e72d1352db4aa9218e928cbe1/endpoints-share.c#L303-L362

Add field Share to Account struct

As denoted by @martensson in
https://github.com/ansd/lastpass-go/issues/43#issuecomment-893537009,
`Group` and `Share` are different things in the LastPass CLI.

The CLI's fullname has the format Shared-folder/groupname/accountname.
For example lpass show --json returns:
    "fullname": "Shared-a/b/c",
    "share": "Shared-a",
    "group": "b",
    "name": "c",

This commit is a breaking change:
In the above example, this library previously set Group to "Shared-a/b".
From now on, this library sets Group to "b" and Share to "Shared-a".

Another breaking change is introduced in the Delete() function.
Instead of just an ID, the function expects *Account because the
function needs to check whether the account is in a shared folder.