Support password iterations != 100100
Closes #44

As reported in https://github.com/lastpass/lastpass-cli/issues/604,
on 18 May 2021 behaviour changed in LastPass servers.

Prior to that date, password iteration != 100100 was supported by this lib by first POSTing
to the /iterations.php endpoint to get the correct iterations count.

This commit fixes the broken /iterations.php behaviour as explained in
https://github.com/detunized/password-manager-access/commit/bd2e31d3478309a2d9a5e171b0422a1a254c3a57#diff-708eab38b171b2961f6da413413fd63d1cff3d5fceda920289959678be35a184R51-R58:

"We no longer request the iteration count from the server in a separate request because it
started to fail in weird ways. It seems there's a special combination of the User Agent and cookies
that returns the correct result. And that is not 100% reliable. After two or three attempts
it starts to fail again with an incorrect result.

So we just went back a few years to the original way LastPass used to handle the iterations.
Namely, submit the default value and if it fails, the error would contain the correct value:
<response><error iterations="5000" /></response>"

So, we first try to login with the default 100100 iterations.
If it fails, we try to login again with the iterations from the error
message.