Allow LastPass sharing RSA private key to be empty
When the user creates an account via
https://lastpass.com/create-account.php
but has never logged in with the LastPass Chrome extension,
an empty privatekeyenc will be returned even if includeprivatekeyenc
is set.

Before this commit, NewClient() failed with the following error:
"input is only 0 bytes; expected at least 16 bytes"

From now on, if privatekeyenc is empty in the HTTP response,
session.privateKey remains nil.
This should be okay because when trying to share a password with an
account that never logged in, we get the following error message in
Chrome:
"Something went wrong sharing with these users:
The user must login to LastPass at least once to permit sharing."

See #52 and
https://github.com/lastpass/lastpass-cli/blob/8767b5e53192ad4e72d1352db4aa9218e928cbe1/cipher.c#L400-L401