Ensure HTTP response body gets closed
"If the returned error is nil, the Response will contain
a non-nil Body which the user is expected to close."
https://pkg.go.dev/net/http#Client.Do

Ensure HTTP response body gets always closed.
Check for errors if closing is not deferred.
Technically, we should always check for errors (i.e. also when
deferred).
However, given the response body is not writable (it is a
io.ReadCloser) it should be good enough to not check for
errors when closing is deferred.